Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.1.2
rq==1.15.1
redis==5.0.1
python-decouple==3.8
PyMuPDF==1.23.18
Pillow==10.2.0
python-docx==0.8.11
pdf2image==1.16.3
google-generativeai==0.8.3